+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 0.1.1
+
+
+
+##### Fixes
+
+- `Orator` class no longer inherits from  `DatabaseManager` to avoid errors when setting up teardown context.
+
+##### Improvements
+
+- Automatically disconnects after a request.
+
+
 ### 0.1
 
 (July 7th, 2015)